requirements.txt
langchain
chromadb
openai
sqlalchemy
python-dotenv
fastapi
uvicorn
langchain_community
sentence-transformers